Reenabling go tests,fixing e2e tests
