+++ Makefile
@@ -1,6 +1,8 @@
 RM = rm -f
 targets=bucky  bucky-fill  bucky-isempty  bucky-pickle-relay  bucky-sparsify  buckyd  findhash  gentestmetrics
 
+OS := $(shell uname)
+
 PHONY: all
 
 all: $(targets)
@@ -29,39 +31,46 @@ findhash:
 gentestmetrics:
 	go build -mod vendor ./cmd/$@
 
+test:
+	pushd ./testing/data && python3 gentests.py && popd
+	go test ./...
+
 # To keep testdata after running test for debugging, run the following command:
 #
-# 	make test REBALANCE_FLAGS=-keep-testdata
+# 	make e2e_test REBALANCE_FLAGS=-keep-testdata
 #
 # Run test_setup_$os first
-test: clean bucky buckyd
-	go run -mod vendor testing/copy.go $(COPY_FLAGS)
-	go run -mod vendor testing/rebalance.go $(REBALANCE_FLAGS)
-	go run -mod vendor testing/backfill2.go $(BACKFILL_FLAGS)
+e2e_test: clean bucky buckyd
+	go run -mod vendor testing/copy/main.go $(COPY_FLAGS)
+	go run -mod vendor testing/rebalance/main.go $(REBALANCE_FLAGS)
+	go run -mod vendor testing/backfill2/main.go $(BACKFILL_FLAGS)
 
-# only works on linux
-test_setup_linux:
+e2e_test_rebalance_health_check: clean bucky buckyd
+	go run -mod vendor testing/rebalance_health_check/main.go $(REBALANCE_FLAGS)
+
+e2e_test_setup:
+ifeq ($(OS),Linux)
 	sudo ip addr add 10.0.1.7 dev lo
 	sudo ip addr add 10.0.1.8 dev lo
 	sudo ip addr add 10.0.1.9 dev lo
-
-test_setup_clean_linux:
-	sudo ip addr del 10.0.1.7/32 dev lo
-	sudo ip addr del 10.0.1.8/32 dev lo
-	sudo ip addr del 10.0.1.9/32 dev lo
-
-test_setup_osx:
+endif
+ifeq ($(OS),Darwin)
 	sudo ifconfig lo0 alias 10.0.1.7 up
 	sudo ifconfig lo0 alias 10.0.1.8 up
 	sudo ifconfig lo0 alias 10.0.1.9 up
+endif
 
-test_setup_clean_osx:
+e2e_test_setup_clean:
+ifeq ($(OS),Linux)
+	sudo ip addr del 10.0.1.7/32 dev lo
+	sudo ip addr del 10.0.1.8/32 dev lo
+	sudo ip addr del 10.0.1.9/32 dev lo
+endif
+ifeq ($(OS),Darwin)
 	sudo ifconfig lo0 -alias 10.0.1.7
 	sudo ifconfig lo0 -alias 10.0.1.8
 	sudo ifconfig lo0 -alias 10.0.1.9
-
-test_rebalance_health_check: clean bucky buckyd
-	go run -mod vendor testing/rebalance_health_check.go $(REBALANCE_FLAGS)
+endif
 
 clean_test:
 	rm -rf bucky buckyd
@@ -69,6 +78,11 @@ clean_test:
 	rm -rf testdata_copy_*
 	rm -rf testdata_backfill2_*
 	rm -rf testdata_rebalance_health_check_*
+	rm -rf testing/data/*.line
+	rm -rf testing/data/*.pickle
+	rm -rf fill/*.wsp
+	rm -rf cmd/bucky-sparsify/sourcefile
+	rm -rf cmd/bucky-sparsify/sourcefile.sparse
 
-clean:
+clean: clean_test
 	$(RM) $(targets)
