+++ cmd/bucky-pickle-relay/main_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"strings"
 	"testing"
 )
@@ -21,19 +20,16 @@ func readfile(s string) ([]byte, error) {
 }
 
 func TestUnMarshal(t *testing.T) {
-	cmd := exec.Command("testdata/gentests.py")
-	err := cmd.Run()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	files, err := ioutil.ReadDir(".")
+	/* test command moded to ./testing/data/gentests.py */
+	/* you need to pre-generate test files in testing/data first */
+	testDatadir := "../../testing/data"
+	testfiles, err := ioutil.ReadDir(testDatadir)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	testcases := 0
-	for _, file := range files {
+	for _, file := range testfiles {
 		if !strings.HasSuffix(file.Name(), ".pickle") {
 			continue
 		}
@@ -41,12 +37,12 @@ func TestUnMarshal(t *testing.T) {
 		testcases++
 		base := file.Name()[:len(file.Name())-7]
 		t.Logf("Testing %s", base)
-		pickle, err := readfile(fmt.Sprintf("%s.pickle", base))
+		pickle, err := readfile(fmt.Sprintf("%s/%s.pickle", testDatadir, base))
 		if err != nil {
 			t.Error(err)
 			continue
 		}
-		blob, err := readfile(fmt.Sprintf("%s.line", base))
+		blob, err := readfile(fmt.Sprintf("%s/%s.line", testDatadir, base))
 		if err != nil {
 			t.Error(err)
 			continue
@@ -82,6 +78,6 @@ func TestUnMarshal(t *testing.T) {
 	}
 
 	if testcases == 0 {
-		t.Fatal("No testcases generated by testdata/gentests.py")
+		t.Fatal("No testcases generated by testing/data/gentests.py")
 	}
 }
